Move Opus header and packet utils into one module
